LeaveType Module Authorized !!

+++ app/Http/Controllers/LeaveTypeController.php
@@ -10,63 +10,82 @@ class LeaveTypeController extends Controller
 
     public function index()
     {
-        $leaveTypes = LeaveType::all();
-        return view('hrm.leavetype.index1', compact('leaveTypes'));
+        if (auth()->user()->can('leavetype_view-all')) {
+            $leaveTypes = LeaveType::all();
+            return view('hrm.leavetype.index1', compact('leaveTypes'));
+        }
+        return abort('403', __('You are not authorized'));
     }
 
 
     public function create()
     {
-        return view('hrm.leavetype.create');
+        if (auth()->user()->can('leavetype_create')) {
+            return view('hrm.leavetype.create');
+        }
+        return abort('403', __('You are not authorized'));
     }
 
 
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
-            'type' => 'required|string|max:200',
-        ]);
-        LeaveType::create($validatedData);
-        return redirect()->route('leaveType.index')->with('success', 'Leave type created successfully');
+        if (auth()->user()->can('leavetype_create')) {
+            $validatedData = $request->validate([
+                'type' => 'required|string|max:200',
+            ]);
+            LeaveType::create($validatedData);
+            return redirect()->route('leaveType.index')->with('success', 'Leave type created successfully');
+        }
+        return abort('403', __('You are not authorized'));
     }
 
     // LeaveTypeController.php
 
     public function edit($id)
     {
-        $leaveType = LeaveType::find($id);
-        
-
-        return view('hrm.leavetype.edit', compact('leaveType'));
+        if (auth()->user()->can('leavetype_edit')) {
+            $leaveType = LeaveType::find($id);
+            return view('hrm.leavetype.edit', compact('leaveType'));
+        }
+        return abort('403', __('You are not authorized'));
     }
 
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'type' => 'required|string|max:150',
-            // Add validation rules for other fields as needed
-        ]);
+        if (auth()->user()->can('leavetype_edit')) {
 
-        $leaveType = LeaveType::find($id);
-        $leaveType->update($request->all());
+            $request->validate([
+                'type' => 'required|string|max:150',
+                // Add validation rules for other fields as needed
+            ]);
 
-        return redirect()->route('leaveType.index')->with('success', 'Leave type updated successfully');
+            $leaveType = LeaveType::find($id);
+            $leaveType->update($request->all());
+
+            return redirect()->route('leaveType.index')->with('success', 'Leave type updated successfully');
+        }
+        return abort('403', __('You are not authorized'));
     }
     public function deleteLeaveType(Request $request)
     {
-        $leaveType = LeaveType::findOrFail($request->id);
-        if($leaveType){
-            $leaveType->delete();
-            return response()->json(['message' => 'Leave Type deleted successfully'], 200);
-        }else{
-            return response()->json(['message' => 'Leave Type not found'], 404);
+        if (auth()->user()->can('leavetype_delete')) {
+            $leaveType = LeaveType::findOrFail($request->id);
+            if ($leaveType) {
+                $leaveType->delete();
+                return response()->json(['message' => 'Leave Type deleted successfully'], 200);
+            } else {
+                return response()->json(['message' => 'Leave Type not found'], 404);
+            }
         }
+        return abort('403', __('You are not authorized'));
     }
 
     public function getData()
     {
-        $leaveTypes = LeaveType::all();
-        return response()->json(['data' => $leaveTypes]);
+        if (auth()->user()->can('leavetype_view-all')) {
+            $leaveTypes = LeaveType::all();
+            return response()->json(['data' => $leaveTypes]);
+        }
+        return abort('403', __('You are not authorized'));
     }
-    
 }
